pyproject.toml
[tool.poetry]
name = "to-do-365-a"
version = "0.1.0"
description = ""
authors = ["Gabriel-Aguiar"]
readme = "README.md"

[tool.poetry.dependencies]
django = "^4.2.7"
djangorestframework = "^3.14.0"
python = "^3.12"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
mkdocs-material = "^9.4.8"
pip-audit = "^2.6.1"
pymdown-extensions = "^10.3.1"
pytest = "^7.4.3"
pylint = "^3.0.2"
mkdocs = "^1.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
